index.js: Configurable reporter options for Formatter
06efe962 (Fix xunit reporter, 2015-04-25) added a hard-coded empty
object for the reporter options.  This commit allows the user to
override that with their own config, falling back to an empty object
if the given value is falsy (for backwards compat with existing
Formatter(...) callers).

If Formatter wasn't already taking a runtime-options argument, I would
have preferred:

  function Formatter (type, options) {
    ...
    var runner = this.runner = new Runner((options || {}).runner)
    this.reporter = new reporters[type](
      this.runner,
      (options || {}).reporter || {},
    )
    Writable.call(this, (options || {}).runner)
    ...
  }

but we can't add 'runner' namespacing to the existing options argument
without breaking backwards compat or adding brittle heuristic
translation.  So I'm just adding a new option, where the reporter is
namespaced to allow for other configurable aspects to also use the new
argument (if we grow other configurable aspects in the future).